// src/app/models/user.ts
export interface ProfileUser {
  uid: string;
  email?: string;
  displayName?: string;
  mobileno?: string;
  address?: string;
  photoURL?: string;
}

export interface User {
  uid: string;
  email: string;
  photoURL?: string;
  displayName?: string;
}
export interface Muser {
  uid?: any;
  email?: string;
  displayName?: string;
  photoURL?:string,
}

//****new */